Update Styles
Add asthetic spacing to icons
