+++ src/index.html
@@ -8,7 +8,10 @@
 <body>
     <div class="container">
       <div class="title-container">
-        <h1 class="app-title">Today's To Do</h1>
+        <div class="title-and-date">
+          <h1 class="app-title">Today's To Do</h1>
+          <div id="date"></div>
+        </div>
         <img class="refresh-icon" src="https://img.icons8.com/external-kmg-design-detailed-outline-kmg-design/64/000000/external-refresh-arrow-kmg-design-detailed-outline-kmg-design.png" alt="Refresh Symbol">
       </div>
       <form class="input-container">
