+++ src/modules/date.js
@@ -1,0 +1,3 @@
+export const dateElement = document.getElementById('date');
+export const options = { weekday: 'long', month: 'short', day: 'numeric' };
+export const today = new Date();
