+++ src/style.css
@@ -8,6 +8,10 @@
   font-size: 16px;
 }
 
+#date {
+  padding-left: 19px;
+}
+
 h1 {
   font-size: 22px;
 }
@@ -40,6 +44,10 @@ button {
   justify-content: space-between;
   height: 2.5rem;
   border-bottom: 3px solid #f6f6f6;
+}
+
+.title-container {
+  padding-bottom: 5px;
 }
 
 .app-title {
@@ -103,6 +111,8 @@ input {
 
 .edit {
   cursor: pointer;
+  padding-bottom: 2px;
+  padding-left: 2px;
 }
 
 .complete span,
